chore(storage-creds): refactor into separate files and use flask_restful for storage provider endpoints
