+++ tests/conftest.py
@@ -363,7 +363,7 @@ def patch_app_db_session(app, monkeypatch):
         monkeypatch.setattr(app.db, 'Session', lambda: session)
         modules_to_patch = [
             'fence.auth',
-            'fence.blueprints.storage_creds',
+            'fence.blueprints.storage_creds.google',
             'fence.oidc.jwt_generator',
             'fence.user',
         ]
@@ -468,7 +468,7 @@ def google_proxy_group(app, db_session, user_client):
 @pytest.fixture(scope='function')
 def cloud_manager():
     manager = MagicMock()
-    patch('fence.blueprints.storage_creds.GoogleCloudManager', manager).start()
+    patch('fence.blueprints.storage_creds.google.GoogleCloudManager', manager).start()
     return manager
